requirements.txt
fake-bpy-module-4.2
numpy
